Collision system... first update
Isn't finished yet. But I decide to stop here to take a breath and to
review what I’ve doing until now. First, this approach of "collision
system", bases on the principle of a matrix from the whole game screen.
Every object that we decide to make “solid” owns a space on the
collision system radar. Every time a "solid" object moves, the
collision system takes it’s coordinates and updates its location
matrix. Later, we'll discuss how to determine when a collision occurs
based on compare a determinate object with the rest of them on radar.